Bucket authorizations

+++ app/controllers/buckets_controller.ts
@@ -1,3 +1,4 @@
+import { bucketAccess } from "#abilities/main";
 import Bucket from "#models/bucket";
 import Category from "#models/category";
 import Item from "#models/item";
@@ -7,10 +8,11 @@ import type { HttpContext } from '@adonisjs/core/http'
 
 export default class BucketsController {
     async list(ctx: HttpContext) {
-        const buckets = await Bucket.query().preload('category');
+        const user = ctx.auth.getUserOrFail();
+        await user.load('buckets', q => q.preload('category'));
 
         return ctx.view.render('pages/buckets/list', {
-            buckets: buckets,
+            buckets: user.buckets,
         });
     }
 
@@ -24,13 +26,22 @@ export default class BucketsController {
 
     async store(ctx: HttpContext) {
         const bucket = await ctx.request.validateUsing(createBucketValidator);
-        await Bucket.create(bucket);
+        await Bucket.create({
+            ...bucket,
+            ownerId: ctx.auth.getUserOrFail().id,
+        });
+
         return ctx.response.redirect().toRoute('bucket_list');
     }
 
     async view(ctx: HttpContext) {
         const id = ctx.params.id;
         const bucket = await Bucket.findOrFail(id);
+
+        if (await ctx.bouncer.denies(bucketAccess, bucket)) {
+            return ctx.response.abort('Your cannot access this bucket', 403)
+        }
+
         await bucket.load('category');
         await bucket.load('items');
 
@@ -42,6 +53,11 @@ export default class BucketsController {
     async delete(ctx: HttpContext) {
         const id = ctx.params.id;
         const bucket = await Bucket.findOrFail(id);
+
+        if (await ctx.bouncer.denies(bucketAccess, bucket)) {
+            return ctx.response.abort('Your cannot access this bucket', 403)
+        }
+
         await bucket.delete();
 
         return ctx.response.redirect().toRoute('bucket_list');
@@ -50,6 +66,10 @@ export default class BucketsController {
     async itemNew(ctx: HttpContext) {
         const id = ctx.params.id;
         const bucket = await Bucket.findOrFail(id);
+
+        if (await ctx.bouncer.denies(bucketAccess, bucket)) {
+            return ctx.response.abort('Your cannot access this bucket', 403)
+        }
 
         const item = await ctx.request.validateUsing(createItemValidator);
