+++ app/controllers/items_controller.ts
@@ -1,3 +1,4 @@
+import { bucketAccess } from "#abilities/main";
 import Item from "#models/item";
 import type { HttpContext } from '@adonisjs/core/http'
 
@@ -5,6 +6,11 @@ export default class ItemsController {
     async done(ctx: HttpContext) {
         const id = ctx.params.id;
         const item = await Item.findOrFail(id);
+        await item.load('bucket');
+
+        if (await ctx.bouncer.denies(bucketAccess, item.bucket)) {
+            return ctx.response.abort('Your cannot access this bucket', 403)
+        }
 
         item.done = !item.done;
         await item.save();
@@ -17,6 +23,11 @@ export default class ItemsController {
     async delete(ctx: HttpContext) {
         const id = ctx.params.id;
         const item = await Item.findOrFail(id);
+
+        if (await ctx.bouncer.denies(bucketAccess, item.bucket)) {
+            return ctx.response.abort('Your cannot access this bucket', 403)
+        }
+
         await item.delete();
 
         return ctx.response.redirect().toRoute('bucket_view', {
